Add common library as a shared common
The project already contains 4 binaries.
HOST device:
- deamon.out
- health-monitor.out

IoT:
- data-provider-battery.out
- data-provider-disk.out

Sometimes the code will be the same with different config
or different parameters passed to the function / class / method.
To achieve this the common directory has been introduced.

Hovewer, sometimes the compile time option like defined (-D)
will be needed (logger is a perfect example).
To make it possible, each binary needs to recompile the common code.
This commit has an example for such a situation.